:truck: chore(api) move `Campaign-participation-route` to `src/devcomp/`
