+++ api/lib/routes.js
@@ -1,7 +1,6 @@
 import * as healthcheck from '../src/shared/application/healthcheck/index.js';
-import * as campaignParticipations from './application/campaign-participations/index.js';
 import * as scoOrganizationLearners from './application/sco-organization-learners/index.js';
 
-const routes = [campaignParticipations, healthcheck, scoOrganizationLearners];
+const routes = [healthcheck, scoOrganizationLearners];
 
 export { routes };
