+++ api/src/devcomp/application/campaign-participations/campaign-participation-route.js
@@ -1,6 +1,6 @@
 import Joi from 'joi';
 
-import { identifiersType } from '../../../src/shared/domain/types/identifiers-type.js';
+import { identifiersType } from '../../../shared/domain/types/identifiers-type.js';
 import { campaignParticipationController } from './campaign-participation-controller.js';
 
 const register = async function (server) {
