+++ api/src/devcomp/routes.js
@@ -1,3 +1,4 @@
+import * as campaignParticipationRoute from './application/campaign-participations/campaign-participation-route.js';
 import * as modulesRoutes from './application/modules/index.js';
 import * as passageEvents from './application/passage-events/index.js';
 import * as passages from './application/passages/index.js';
@@ -7,6 +8,7 @@ import * as userTrainings from './application/user-trainings/index.js';
 import * as userTutorials from './application/user-tutorials/index.js';
 
 const devcompRoutes = [
+  campaignParticipationRoute,
   modulesRoutes,
   passages,
   passageEvents,
