+++ api/tests/devcomp/acceptance/application/campaign-participations/campaign-participation-controller_test.js
@@ -3,7 +3,7 @@ import {
   databaseBuilder,
   expect,
   generateAuthenticatedUserRequestHeaders,
-} from '../../../test-helper.js';
+} from '../../../../test-helper.js';
 
 describe('Acceptance | API | Campaign Participations', function () {
   let server, user;
