+++ api/tests/unit/application/campaign-participations/campaign-participation-controller_test.js
@@ -1,4 +1,4 @@
-import { campaignParticipationController } from '../../../../lib/application/campaign-participations/campaign-participation-controller.js';
+import { campaignParticipationController } from '../../../../src/devcomp/application/campaign-participations/campaign-participation-controller.js';
 import { usecases as devcompUsecases } from '../../../../src/devcomp/domain/usecases/index.js';
 import { expect, sinon } from '../../../test-helper.js';
 describe('Unit | Application | Controller | Campaign-Participation', function () {
