+++ api/tests/unit/application/campaign-participations/index_test.js
@@ -1,5 +1,5 @@
-import { campaignParticipationController } from '../../../../lib/application/campaign-participations/campaign-participation-controller.js';
-import * as moduleUnderTest from '../../../../lib/application/campaign-participations/index.js';
+import { campaignParticipationController } from '../../../../src/devcomp/application/campaign-participations/campaign-participation-controller.js';
+import * as moduleUnderTest from '../../../../src/devcomp/application/campaign-participations/campaign-participation-route.js';
 import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
 
 describe('Unit | Application | Router | campaign-participation-router ', function () {
